+++ R/colnames.R
@@ -15,13 +15,13 @@
 #' check_colnames(data, c("y", "x"), exclusive = TRUE, error = FALSE)
 #' check_colnames(data, c("y", "x"), order = TRUE, error = FALSE)
 #' check_colnames(data, c("a"), error = FALSE)
-check_colnames <- function(x, colnames, exclusive = FALSE, order = FALSE,
+check_colnames <- function(x, colnames = character(0), exclusive = FALSE, order = FALSE,
                          x_name = substitute(x),
                          error = TRUE) {
 
   x_name <- deparse_x_name(x_name)
 
-  check_vector(colnames, "", length = c(1L, .Machine$integer.max), unique = TRUE)
+  check_vector(colnames, "", unique = TRUE)
   check_flag_internal(exclusive)
   check_flag_internal(order)
   check_flag_internal(error)
@@ -30,6 +30,12 @@ check_colnames <- function(x, colnames, exclusive = FALSE, order = FALSE,
   x_colnames <- colnames(x)
   
   if(is.null(x_colnames)) on_fail(x_name, " must have column names", error = error)
+  
+  if(!length(colnames)) {
+    if(exclusive && length(x_colnames))
+        on_fail(x_name, " must not have any columns", error = error)
+    return(x)
+  }
 
   if (exclusive) {
     if (order) {