+++ NEWS.md
@@ -1,3 +1,5 @@
+- check_colnames and check_names now accept (col)names = character(0)
+
 # checkr 0.0.1
 
 - First official release